third_party/WebKit/LayoutTests/fast/block/scrollbar-wider-than-border-box.html

<!DOCTYPE html>
<div style="position:relative;">
    <div style="overflow:scroll; width:1px;">
        <div id="inner" style="float:right; width:20px; height:100px;"></div>
    </div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
    test(() => {
        assert_equals(inner.offsetLeft, -20);
    }, "Right-aligned float in zero-width block");
</script>
